fix: rename productDetail
